backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.35.0
sqlalchemy>=2.0.36,<2.1.0
python-dotenv>=1.0.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<1.8.0
pydantic>=2.7.0,<3.0.0
email-validator>=2.0.0,<3.0.0
python-multipart>=0.0.9,<0.1.0
